backend/requirements.txt
flask
requests
pymongo
google-cloud-dialogflow
flask-cors
gunicorn
google-generativeai
